+++ apps/drec-api/src/drec.module.ts
@@ -70,6 +70,7 @@ import { BulkUploadEntity } from './pods/bulk-upload/bulk-uploads.entity';
 import { BulkUploadFailedLogEntity } from './pods/bulk-upload/bulk-uploads-failed-logs.entity';
 import { BulkUploadModule } from './pods/bulk-upload/bulk-upload.module';
 import { HealthModule } from './pods/health/health.module';
+import { getTypeOrmConfig } from './config/typeorm.config';
 const getEnvFilePath = () => {
   const pathsToTest = [
     '../../../.env',
@@ -124,26 +125,10 @@ export const entities = [
 ];
 
 export const originAppTypeOrmModule = (): DynamicModule => {
-  const options: TypeOrmModuleOptions = process.env.DATABASE_URL
-    ? {
-        type: 'postgres',
-        url: process.env.DATABASE_URL,
-        ssl: {
-          rejectUnauthorized: false,
-        },
-        entities,
-        logging: ['info'],
-      }
-    : {
-        type: 'postgres',
-        host: process.env.DB_HOST ?? 'localhost',
-        port: Number(process.env.DB_PORT) ?? 5432,
-        username: process.env.DB_USERNAME ?? 'postgres',
-        password: process.env.DB_PASSWORD ?? 'postgres',
-        database: process.env.DB_DATABASE ?? 'origin',
-        entities,
-        logging: ['info'],
-      };
+  const options: TypeOrmModuleOptions = {
+    ...getTypeOrmConfig(),
+    entities,
+  };
 
   return TypeOrmModule.forRoot(options);
 };